requirements.txt
# build tools
cmake
ninja

# web tools
PyGithub
requests

# computational libraries
numpy
einops
scipy
kornia
opencv-python
pillow
nerfacc>=0.5.3
slangtorch

# for use ML models
diffusers>=0.26.1
transformers>=4.36.2
safetensors
open_clip_torch

# for training differentiable tensors
torchmetrics
pytorch_msssim
pytorch-lightning
peft

# for process images & videos
rembg
imageio
imageio-ffmpeg
matplotlib

# for process, import & export mesh 
trimesh
plyfile
pygltflib
xatlas
pymeshlab
gpytoolbox
PyMCubes
libigl

# configs & extra
torchtyping>=0.1.5
tqdm
jaxtyping
packaging
OmegaConf
pyhocon
iopath